requirements.txt
pdfplumber>=0.7.6
sentence-transformers>=2.2.2
torch>=1.13.1   # CPU or GPU compatible version
numpy>=1.21.0
pandas>=1.3.0
networkx>=2.6
pyvis>=0.3.0
tqdm>=4.60.0
openai>=0.27.0
python-dotenv>=0.19.0



